Fixed line endings (CLRF) and added kNN graph coloring for 75/25 percentiles.

+++ src/main/java/pl/wroc/pwr/patryk/MyDirectedGraph.java
@@ -81,7 +81,7 @@ public class MyDirectedGraph{
 					//iterate over all values in a row (split by ';')
 					for (String s : nextLine.split(";")) {
 						try {
-							vertex.addValue( Float.parseFloat(s) );
+							vertex.addValue( Double.parseDouble(s) );
 						} catch (NumberFormatException e) {
 							
 							e.printStackTrace();
